Extract justification verifier.
